fix: warning location
